% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fieldCrop_grid.R
\name{fieldCrop_grid}
\alias{fieldCrop_grid}
\title{Cropping image according to the plot grid shapefile.}
\usage{
fieldCrop_grid(
  mosaic,
  fieldShape,
  classifier = NULL,
  plotID = NULL,
  format = ".tif",
  output_dir = "./"
)
}
\arguments{
\item{mosaic}{image object format \code{\link{rast}} or \code{\link{stars}}.}

\item{fieldShape}{crop the image using the fieldShape as reference.}

\item{classifier}{column name on fieldShape with levels to create folders at 'output_dir' directory. Cropped plots will be sorted and stored on related folders. If NULL all plots will be saved at the working directory.}

\item{plotID}{column name on fieldShape with the desired plots identification. If NULL plots will be saved in sequence according to the field map.}

\item{format}{save plots as \code{\link{.tif}} (by default) or \code{".jpg"}.}

\item{output_dir}{directory to save cropped plot images. Default is the working directory.}
}
\value{
A list with plots image format 'stars'. The function also saves plot images identified by 'plotID' sorted according to the 'classifier' folder at the 'output_dir' directory.
}
\description{
It crops each plot and associate it with a ID and store it according to the user specification
}
\details{
fieldCrop_grid
}
