% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fieldKmeans.R
\name{fieldKmeans}
\alias{fieldKmeans}
\title{Clustering pixels using kmeans algorithm.}
\usage{
fieldKmeans(
  mosaic,
  clusters = 3,
  iteration = 500,
  algorithm = "Lloyd",
  plot = TRUE,
  seed = 25
)
}
\arguments{
\item{mosaic}{image object format \code{\link{rast}} or \code{\link{stars}}.}

\item{clusters}{number of clusters to sort pixels (Default = 3).}

\item{iteration}{number of iterations (Default = 500).}

\item{algorithm}{kmeans algorithm. Check for options on \code{\link{stats::kmeans}} (Default = "Lloyd").}

\item{plot}{plot clusters}

\item{seed}{set.seed}
}
\value{
An image object format \code{\link{rast}} identified with clusters.
}
\description{
It allows cluster pixels on k groups that can be related with plants, soil, shadows, etc.
}
\details{
fieldKmeans
}
