% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fieldSegment.R
\name{fieldSegment}
\alias{fieldSegment}
\title{Clustering pixels with Machine Learning models.}
\usage{
fieldSegment(mosaic, trainDataset, model = "rf", seed = 5)
}
\arguments{
\item{mosaic}{image object format \code{\link{rast}} or \code{\link{stars}}.}

\item{trainDataset}{number of clusters to sort pixels (Default = 3).}

\item{model}{machine learning models for classification objects in the mosaic based on samples. There are two options the model Randon Forest ('rf'). (Default='rf')}

\item{seed}{set.seed}
}
\value{
An image object format \code{\link{rast}} identified with clusters.
}
\description{
It allows clustering pixels based on training image samples (n clusters) using machine learning models that can be related with plants, soil, shadows, etc.
}
\details{
fieldSegment
}
