% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fieldShape_edit.R
\name{fieldShape_edit}
\alias{fieldShape_edit}
\title{Editing each plot in the grid of a shapefile.}
\usage{
fieldShape_edit(
  mosaic,
  fieldShape = NULL,
  shp = NULL,
  r = 1,
  g = 2,
  b = 3,
  color_options = viridisLite::viridis,
  max_pixels = 1e+08,
  downsample = 5
)
}
\arguments{
\item{mosaic}{image object format \code{\link{rast}} or \code{\link{stars}}.}

\item{fieldShape}{plot grid 'fieldShape' object to be edited from \code{\link{fieldShape_render()}}.}

\item{shp}{plot grid made in any other GIS software. Please use \code{\link{shp=st_read("path to .shp file")}}.}

\item{r}{red layer in the mosaic (for RGB image normally is 1). If NULL the first single layer will be plotted.}

\item{g}{green layer in the mosaic (for RGB image normally is 2). If NULL the first single layer will be plotted.}

\item{b}{blue layer in the mosaic (for RGB image normally is 3). If NULL the first single layer will be plotted.}

\item{color_options}{single layer coloring options. Check more information at \code{\link{color_options}} from \code{\link{leafem}} package.}

\item{max_pixels}{maximun pixels allowed before down sampling. Reducing size to accelerate analysis. Default = 100000000.}

\item{downsample}{numeric downsample reduction factor. Default = 5.}
}
\value{
A new edited plot 'fieldShape' class "sf" & "data.frame".
}
\description{
It allows editing each plot (change shape, remove, move, etc.).
}
\details{
fieldShape_edit
}
