% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fieldShape_render.R
\name{fieldShape_render}
\alias{fieldShape_render}
\title{Building the plot \code{fieldshape} file using Zoom Visualization}
\usage{
fieldShape_render(
  mosaic,
  ncols,
  nrows,
  fieldData = NULL,
  fieldMap = NULL,
  PlotID = NULL,
  buffer = NULL,
  plot_size = NULL,
  r = 1,
  g = 2,
  b = 3,
  color_options = viridisLite::viridis,
  max_pixels = 1e+08,
  downsample = 5
)
}
\arguments{
\item{mosaic}{object of class 'rast' or 'stars' obtained from function \code{\link{terra::rast()}}.}

\item{ncols}{number of columns.}

\item{nrows}{number of rows.}

\item{fieldData}{data frame with plot ID and all attributes of each plot (Traits as columns and genotypes as rows).}

\item{fieldMap}{matrix with plots ID identified by rows and ranges, please use first the funsction \code{\link{fieldMap}}.}

\item{buffer}{negative values should be used to remove boundaries from neighbor plot.}

\item{plot_size}{specific plot shape size c(x,y). For example, mosaic with pixels in meter \code{\link{plot_size=c(0.5,1.5)}} means 0.5m x 1.5cm the plot shape.}

\item{r}{red layer in the mosaic (for RGB image normally is 1). If NULL the first single layer will be plotted.}

\item{g}{green layer in the mosaic (for RGB image normally is 2). If NULL the first single layer will be plotted.}

\item{b}{blue layer in the mosaic (for RGB image normally is 3). If NULL the first single layer will be plotted.}

\item{color_options}{single layer coloring options. Check more information at \code{\link{color_options}} from \code{\link{leafem}} package.}

\item{max_pixels}{maximun pixels allowed before down sampling. Reducing size to accelerate analysis. Default = 100000000.}

\item{downsample}{numeric downsample reduction factor. Default = 5.}

\item{plotID}{name of plot ID in the fieldData file to combine with fieldShape.}
}
\value{
A field plot shape file class "sf" & "data.frame".
}
\description{
The user should select the four experimental field corners and the shape file with plots will be automatically 
built using a grid with the number of ranges and rows. Attention: The clicking sequence must be (1) upper left, (2) upper right, (3) lower right, and (4) lower left.
}
\details{
fieldShape_render
}
