% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fieldView.R
\name{fieldView}
\alias{fieldView}
\title{Image and plot grid shape file visualization}
\usage{
fieldView(
  mosaic = NULL,
  fieldShape = NULL,
  plotCol = NULL,
  editor = FALSE,
  type = 1,
  r = 1,
  g = 2,
  b = 3,
  colorOptions = viridisLite::viridis,
  col_grid = viridisLite::viridis,
  alpha_grid = 1,
  seq_grid = NULL,
  max_pixels = 1e+08,
  downsample = 5
)
}
\arguments{
\item{mosaic}{object of class 'rast' or 'stars' obtained from function \code{\link{terra::rast()}}.}

\item{fieldShape}{plot grid 'fieldShape' object to be edited from \code{\link{fieldShape_render()}}.}

\item{plotCol}{column name in  'fieldShape' to color each plot.}

\item{editor}{editing image.}

\item{type}{if 1 mosaic and 'fieldShape' will be visualized side by side. If 2, 'mosaic' and 'fieldShape' will be overlayed (Attention: only small size mosaics can be visualized on type=2).}

\item{r}{red layer in the mosaic (for RGB image normally is 1). If NULL the first single layer will be plotted.}

\item{g}{green layer in the mosaic (for RGB image normally is 2). If NULL the first single layer will be plotted.}

\item{b}{blue layer in the mosaic (for RGB image normally is 3). If NULL the first single layer will be plotted.}

\item{colorOptions}{single layer coloring options. Check more information at \code{\link{colorOptions}} from \code{\link{leafem}} package (check for 'palette').}

\item{col_grid}{grid color options similar to 'col.regions' parameter in 'mapview'.}

\item{alpha_grid}{transparency with values between 0 and 1.}

\item{seq_grid}{color classes similar to 'at' parameter in 'mapview'.}

\item{max_pixels}{maximun pixels allowed before down sampling. Reducing size to accelerate analysis. Default = 100000000.}

\item{downsample}{numeric downsample reduction factor. Default = 5.}
}
\value{
Mosaic and fieldShape visualization object.
}
\description{
Graphic visualization of trait values for each plot using the \code{\link{fieldShape}} file and original image.
}
\details{
fieldView
}
